Further unit test suite tweats
